updates per feedback
